pyproject.toml
[project]
name = "zukko-ai"
version = "0.1.0"
description = "Zukko AI Telegram tutor"
requires-python = ">=3.10"
dependencies = [
  "pyTelegramBotAPI>=4.14.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
